Fix transcription error
I had to go back to the original 3d-bin-pack repo and the print off appendix b of the pdf (the c program code), and work my way line-by-line through the code, looking for differences.  Found a single (but major) difference on line 980 of the original transcribed code, which I'm porting over to here.  This fixes #6 (unexpected run lengths), and fixes #7 (free() errors).